server/app/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.14,<2.0
PyJWT>=2.10,<3.0
bcrypt>=5.0.0
pydantic>=2.10,<3.0
pydantic-settings>=2.6,<3.0
email-validator>=2.2,<3.0
aiosqlite>=0.22.1
# System Monitoring (Control Tower)
psutil>=6.0,<7.0

# Payment Processing
stripe>=10.0,<11.0
paypalrestsdk>=1.7,<2.0


# HTTP Client for webhooks and external APIs
requests>=2.31,<3.0
httpx>=0.25,<1.0

# Email delivery with templates
Jinja2>=3.0,<4.0

